third_party/blink/web_tests/platform/mac-mac13-arm64/external/wpt/html/canvas/offscreen/manual/the-offscreen-canvas/offscreencanvas.getcontext-expected.txt

This is a testharness.js-based test.
PASS Test that getContext with un-supported string throws a TypeError.
FAIL Test that getContext with supported string returns correct results assert_true: expected true got false
FAIL Test that getContext twice with different context type returns null the second time assert_equals: expected null but got object "[object OffscreenCanvasRenderingContext2D]"
PASS Test that 2dcontext.canvas should return the original OffscreenCanvas
FAIL Test that webglcontext.canvas should return the original OffscreenCanvas Cannot read properties of null (reading 'canvas')
PASS Test that OffscreenCanvasRenderingContext2D with alpha disabled makes the OffscreenCanvas opaque
PASS Test that OffscreenCanvasRenderingContext2D with alpha enabled preserves the alpha
PASS Test that 'alpha' context creation attribute is true by default
Harness: the test ran to completion.
